+++ tests/conftest.py
@@ -117,6 +117,7 @@ pytest_plugins = [
     "tests.fixture_plugins.analysis_starter.store_fixtures",
     "tests.fixture_plugins.analysis_starter.params_file_creators",
     "tests.fixture_plugins.analysis_starter.params_file_content_fixtures",
+    "tests.fixture_plugins.analysis_starter.testing_scenarios",
     "tests.fixture_plugins.backup_fixtures.backup_fixtures",
     "tests.fixture_plugins.chanjo2_fixtures.api_fixtures",
     "tests.fixture_plugins.chanjo2_fixtures.models_fixtures",
